feat: support persisting IDE UI layout and etc.

+++ install.sh
@@ -1,22 +1,31 @@
 #!/usr/bin/env bash
+set -eu
+
+function install_dotfiles() {
+  local source="${1:-}"
+  local target="${2:-}"
+  if test ! -e "${source:-}"; then return 0; fi
+
+  while read -r file; do
+
+      relative_file_path="${file#"${source}"/}"
+      target_file="${target}/${relative_file_path}"
+      target_dir="${target_file%/*}"
+
+      if test ! -d "${target_dir}"; then
+          mkdir -p "${target_dir}"
+      fi
+
+      printf 'Installing dotfiles symlink %s\n' "${target_file}"
+      ln -sf "${file}" "${target_file}"
+
+  done < <(find "${source}" -type f)
+}
 
 current_dir="$(cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd)"
-dotfiles_source="${current_dir}/home_files"
 
-while read -r file; do
-
-    relative_file_path="${file#"${dotfiles_source}"/}"
-    target_file="${HOME}/${relative_file_path}"
-    target_dir="${target_file%/*}"
-
-    if test ! -d "${target_dir}"; then
-        mkdir -p "${target_dir}"
-    fi
-
-    printf 'Installing dotfiles symlink %s\n' "${target_file}"
-    ln -sf "${file}" "${target_file}"
-
-done < <(find "${dotfiles_source}" -type f)
+install_dotfiles "${current_dir}/home_files" "${HOME}"
+install_dotfiles "${current_dir}/workspace_repo" "${GITPOD_REPO_ROOT}"
 
 if test ! -e /usr/bin/commitjb; then
     sudo ln -s "${current_dir}/commitjb" /usr/bin/
